It works

+++ internal/models/hacker.go
@@ -1,6 +1,6 @@
 package models
 
 type Hacker struct {
-	Name  string `json:"name,omitempty"`
-	Score int    `json:"score,omitempty"`
+	Name  string  `json:"name,omitempty"`
+	Score float64 `json:"score,omitempty"`
 }
